feat(provider): add /neuralwatt:fetch, remove hardcoded model list
Replace the hardcoded public model list and session_start discovery cache
with a single on-demand command. /neuralwatt:fetch hits the authenticated
/v1/models endpoint and atomically replaces the provider's model list via
registerProvider (clear + write in one call).

- models/index.ts: single loadNeuralwattModels() returning all live models;
  public/hidden split, dedupe step, and disk cache removed
- provider/index.ts: ships empty model list, drops session_start discovery,
  registers /neuralwatt:fetch; quota/rate-limit/context-overflow handlers
  unchanged
- config.ts + schema.json: remove inert includeLegacyModelIds and
  includeHiddenModels toggles
- delete models/public-models.ts, legacy.ts, hidden.ts, cache.ts

Behavior change: provider starts with no models each session; run
/neuralwatt:fetch to populate the picker.
