+++ schema.json
@@ -5,14 +5,6 @@
     "NeuralwattConfig": {
       "additionalProperties": false,
       "properties": {
-        "includeHiddenModels": {
-          "description": "Include hidden Neuralwatt models discovered via the authenticated API.",
-          "type": "boolean"
-        },
-        "includeLegacyModelIds": {
-          "description": "Include legacy Neuralwatt model IDs in the model picker.",
-          "type": "boolean"
-        },
         "quotaCommand": {
           "description": "Show the quota command (/neuralwatt:quota).",
           "type": "boolean"
