+++ src/config.ts
@@ -4,7 +4,6 @@ import {
   type SettingsSection,
 } from "@aliou/pi-utils-settings";
 import type { ExtensionAPI } from "@earendil-works/pi-coding-agent";
-import type { SettingItem } from "@earendil-works/pi-tui";
 
 export type NeuralwattFeatureId =
   | "quotaCommand"
@@ -28,42 +27,24 @@ export interface NeuralwattConfig {
   quotaWarnings?: boolean;
   /** Show usage in the sub-bar / status bar. */
   subBarIntegration?: boolean;
-  /** Include legacy Neuralwatt model IDs in the model picker. */
-  includeLegacyModelIds?: boolean;
-  /** Include hidden Neuralwatt models discovered via the authenticated API. */
-  includeHiddenModels?: boolean;
 }
 
 export interface ResolvedNeuralwattConfig {
   quotaCommand: boolean;
   quotaWarnings: boolean;
   subBarIntegration: boolean;
-  includeLegacyModelIds: boolean;
-  includeHiddenModels: boolean;
 }
 
 const DEFAULTS: ResolvedNeuralwattConfig = {
   quotaCommand: true,
   quotaWarnings: true,
   subBarIntegration: true,
-  includeLegacyModelIds: false,
-  includeHiddenModels: false,
 };
 
 export const configLoader = new ConfigLoader<
   NeuralwattConfig,
   ResolvedNeuralwattConfig
->("neuralwatt", DEFAULTS, {
-  migrations: [
-    {
-      name: "disable-legacy-model-ids-by-default",
-      shouldRun: (config) => config.includeLegacyModelIds === undefined,
-      message:
-        "[neuralwatt] legacy model IDs (ids including the provider and the quantization) are disabled by default. You can enable them with /neuralwatt:settings.",
-      run: (config) => ({ ...config, includeLegacyModelIds: false }),
-    },
-  ],
-});
+>("neuralwatt", DEFAULTS);
 
 export const NEURALWATT_CONFIG_UPDATED_EVENT =
   "neuralwatt:config:updated" as const;
@@ -88,7 +69,7 @@ function featureRow(
   description: string,
   configValue: boolean,
   isLoaded: boolean,
-): SettingItem {
+) {
   if (isLoaded) {
     return {
       id,
@@ -146,47 +127,9 @@ export function registerNeuralwattSettings(
             ),
           ],
         },
-        {
-          label: "Other settings",
-          items: [
-            {
-              id: "includeLegacyModelIds",
-              label: "Legacy model IDs",
-              description:
-                "Include deprecated Neuralwatt model IDs as aliases in the model picker",
-              currentValue:
-                (tabConfig?.includeLegacyModelIds ??
-                resolved.includeLegacyModelIds)
-                  ? "include"
-                  : "ignore",
-              values: ["include", "ignore"],
-            },
-            {
-              id: "includeHiddenModels",
-              label: "Hidden models",
-              description:
-                "Include Neuralwatt models that are accessible via API key but not advertised in the public model list",
-              currentValue:
-                (tabConfig?.includeHiddenModels ?? resolved.includeHiddenModels)
-                  ? "include"
-                  : "ignore",
-              values: ["include", "ignore"],
-            },
-          ],
-        },
       ];
     },
     onSettingChange: (id, newValue, config) => {
-      // Non-feature toggles are handled first so they are not blocked by the
-      // loaded-features guard (they are managed directly by the provider).
-      if (id === "includeLegacyModelIds") {
-        return { ...config, includeLegacyModelIds: newValue === "include" };
-      }
-
-      if (id === "includeHiddenModels") {
-        return { ...config, includeHiddenModels: newValue === "include" };
-      }
-
       if (!getLoadedFeatures().has(id as NeuralwattFeatureId)) {
         return null;
       }
